chore: support both simple and payment form on the demo
