+++ demo/.env.template
@@ -1,0 +1,3 @@
+VITE_SIMPLE_FORM_ID=
+VITE_PAYMENT_FORM_ID=
+VITE_STRIPE_PUBLISHABLE_KEY=
