+++ demo/package.json
@@ -16,7 +16,7 @@
   "devDependencies": {
     "@types/react": "^16.9.11",
     "@types/react-dom": "^16.8.4",
-    "typescript": "^4",
+    "typescript": "^4.7.4",
     "vite": "latest",
     "vite-preset-react": "latest"
   },
